+++ model/star.go
@@ -115,35 +115,97 @@ func FindStarByID(db *gorm.DB, ID int) (Star, error) {
 }
 
 // FindStars finds all stars
-func FindStars(db *gorm.DB) ([]Star, error) {
+func FindStars(db *gorm.DB, match string) ([]Star, error) {
 	var stars []Star
-	db.Order("full_name").Find(&stars)
+	if match != "" {
+		db.Where("full_name LIKE ?",
+			strings.ToLower(fmt.Sprintf("%%%s%%", match))).Order("full_name").Find(&stars)
+	} else {
+		db.Order("full_name").Find(&stars)
+	}
 	return stars, db.Error
 }
 
 // FindUntaggedStars finds stars without any tags
-func FindUntaggedStars(db *gorm.DB) ([]Star, error) {
-    var stars []Star
-    db.Raw("SELECT * FROM STARS S WHERE S.ID NOT IN (SELECT STAR_ID FROM STAR_TAGS)").Scan(&stars)
-    return stars, db.Error
+func FindUntaggedStars(db *gorm.DB, match string) ([]Star, error) {
+	var stars []Star
+	if match != "" {
+		db.Raw(`
+			SELECT *
+			FROM STARS S
+			WHERE S.DELETED_AT IS NULL
+			AND S.FULL_NAME LIKE ?
+			AND S.ID NOT IN (
+				SELECT STAR_ID
+				FROM STAR_TAGS
+			) ORDER BY S.FULL_NAME`,
+			fmt.Sprintf("%%%s%%", strings.ToLower(match))).Scan(&stars)
+	} else {
+		db.Raw(`
+			SELECT *
+			FROM STARS S
+			WHERE S.DELETED_AT IS NULL
+			AND S.ID NOT IN (
+				SELECT STAR_ID
+				FROM STAR_TAGS
+			) ORDER BY S.FULL_NAME`).Scan(&stars)
+	}
+	return stars, db.Error
 }
 
 // FindStarsByLanguageAndOrTag finds stars with the specified language and/or the specified tag
-func FindStarsByLanguageAndOrTag(db *gorm.DB, language string, tagName string, union bool) ([]Star, error) {
+func FindStarsByLanguageAndOrTag(db *gorm.DB, match string, language string, tagName string, union bool) ([]Star, error) {
 	operator := "AND"
 	if union {
 		operator = "OR"
 	}
 
 	var stars []Star
-	db.Raw("SELECT * FROM STARS S, TAGS T INNER JOIN STAR_TAGS ST ON S.ID = ST.STAR_ID INNER JOIN TAGS ON ST.TAG_ID = T.ID WHERE LOWER(T.NAME) = ? "+operator+" LOWER(S.LANGUAGE) = ? GROUP BY ST.STAR_ID", strings.ToLower(tagName), strings.ToLower(language)).Scan(&stars)
+	if match != "" {
+		db.Raw(fmt.Sprintf(`
+			SELECT * 
+			FROM STARS S, TAGS T 
+			INNER JOIN STAR_TAGS ST ON S.ID = ST.STAR_ID 
+			INNER JOIN TAGS ON ST.TAG_ID = T.ID 
+			WHERE S.DELETED_AT IS NULL
+			AND T.DELETED_AT IS NULL
+			AND LOWER(S.FULL_NAME) LIKE ? 
+			AND (LOWER(T.NAME) = ? 
+			%s LOWER(S.LANGUAGE) = ?) 
+			GROUP BY ST.STAR_ID 
+			ORDER BY S.FULL_NAME`, operator),
+			fmt.Sprintf("%%%s%%", strings.ToLower(match)),
+			strings.ToLower(tagName),
+			strings.ToLower(language)).Scan(&stars)
+	} else {
+		db.Raw(fmt.Sprintf(`
+			SELECT * 
+			FROM STARS S, TAGS T 
+			INNER JOIN STAR_TAGS ST ON S.ID = ST.STAR_ID 
+			INNER JOIN TAGS ON ST.TAG_ID = T.ID 
+			WHERE S.DELETED_AT IS NULL
+			AND T.DELETED_AT IS NULL
+			AND LOWER(T.NAME) = ? 
+			%s LOWER(S.LANGUAGE) = ? 
+			GROUP BY ST.STAR_ID 
+			ORDER BY S.FULL_NAME`, operator),
+			strings.ToLower(tagName),
+			strings.ToLower(language)).Scan(&stars)
+	}
 	return stars, db.Error
 }
 
 // FindStarsByLanguage finds stars with the specified language
-func FindStarsByLanguage(db *gorm.DB, language string) ([]Star, error) {
+func FindStarsByLanguage(db *gorm.DB, match string, language string) ([]Star, error) {
 	var stars []Star
-	db.Where("lower(language) = ?", strings.ToLower(language)).Order("full_name").Find(&stars)
+	if match != "" {
+		db.Where("full_name LIKE ? AND lower(language) = ?",
+			strings.ToLower(fmt.Sprintf("%%%s%%", match)),
+			strings.ToLower(language)).Order("full_name").Find(&stars)
+	} else {
+		db.Where("lower(language) = ?",
+			strings.ToLower(language)).Order("full_name").Find(&stars)
+	}
 	return stars, db.Error
 }